# squash headers into single segment during parsing