--- a/opte/src/engine/snat.rs
+++ b/opte/src/engine/snat.rs
@@ -220,8 +220,9 @@ impl SNat {
             if icmp.msg_type() != Icmpv4Message::EchoRequest {
                 return Err(GenDescError::Unexpected {
                     msg: format!(
-                        "Expected ICMP Echo Request, found: {}",
-                        icmp.msg_type()
+                        "Expected ICMP Echo Request, found: {} \n{:x?}",
+                        icmp.msg_type(),
+                        body_segs,
                     ),
                 });
             }