--- a/oxide-vpc/tests/common/icmp.rs
+++ b/oxide-vpc/tests/common/icmp.rs
@@ -123,16 +123,21 @@ pub fn gen_icmp_echo(
         eth.emit(wtr.slice_mut(EtherHdr::SIZE).unwrap());
 
         /*
-        let i = pkt.add_seg(ip4.hdr_len());
-        let mut wtr = pkt.seg_wtr(i);
-        ip4.emit(wtr.slice_mut(ip4.hdr_len()).unwrap());
+         * TODO
+         * It seems like this should also work, but it does not. Leaving this as
+         * a todo, as I've not actually seen packets like this in the wild yet.
+         *
+         *   let i = pkt.add_seg(ip4.hdr_len()).unwrap();
+         *   let mut wtr = pkt.seg_wtr(i);
+         *   ip4.emit(wtr.slice_mut(ip4.hdr_len()).unwrap());
+         *
+         *   let i = pkt.add_seg(icmp_bytes.len()).unwrap();
+         *   let mut wtr = pkt.seg_wtr(i);
+         *   wtr.write(&icmp_bytes).unwrap();
+         *
+         */
 
-        let i = pkt.add_seg(icmp_bytes.len());
-        let mut wtr = pkt.seg_wtr(i);
-        wtr.write(&icmp_bytes).unwrap();
-        */
-
-        let i = pkt.add_seg(ip4.hdr_len() + icmp_bytes.len());
+        let i = pkt.add_seg(ip4.hdr_len() + icmp_bytes.len()).unwrap();
         let mut wtr = pkt.seg_wtr(i);
         ip4.emit(wtr.slice_mut(ip4.hdr_len()).unwrap());
         wtr.write(&icmp_bytes).unwrap();
@@ -181,6 +186,13 @@ pub fn gen_icmpv6_echo_req(
         wtr.write(&body_bytes).unwrap();
         pkt.parse(Out, VpcParser::new()).unwrap()
     } else {
-        todo!();
+        let mut pkt = Packet::alloc_and_expand(EtherHdr::SIZE);
+        let mut wtr = pkt.seg_wtr(0);
+        eth.emit(wtr.slice_mut(EtherHdr::SIZE).unwrap());
+        let i = pkt.add_seg(ip6.hdr_len() + body_bytes.len()).unwrap();
+        let mut wtr = pkt.seg_wtr(i);
+        ip6.emit(wtr.slice_mut(ip6.hdr_len()).unwrap());
+        wtr.write(&body_bytes).unwrap();
+        pkt.parse(Out, VpcParser::new()).unwrap()
     }
 }