This PR fixes the issue described in the following [comment](https://github.com/oxidecomputer/opte/blob/master/opte/src/engine/packet.rs#L1292-L1307). 

```
XXX This assumes that the headers all reside in the
first segment. For any typical implementation, this
should be true (it's better for performance, and
just makes the most sense). However, if that
invariant doesn't hold true, this method of erasing
the original header data is incomplete. It will
only partially erase the data, leading to confusion
downstream somewhere. The best solution is to check
for this during parsing. If the header straddles
segments, then just copy all header data into a new
segment, and replace the old header data with that
one segment. This means a hit on performance, but
it also means sanity for all downstream code.
Besides, any network stack that cares about
performance will already make sure to place the
headers in a single buffer.
```

This has resulted in panics in a few different environments.

This patch takes the advice of the comment and squashes the headers down into a single segment during parsing.

## Testing

A test is also added that captures the panic behavior.

### Without patch

```
ry@masaka:~/src/opte/oxide-vpc$ RUST_BACKTRACE=1 cargo test snat_icmp4_echo_rewrite
    Finished test [unoptimized + debuginfo] target(s) in 0.07s
     Running unittests src/lib.rs (target/debug/deps/oxide_vpc-09fa660a95f8ab19)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 6 filtered out; finished in 0.00s

     Running tests/firewall_tests.rs (target/debug/deps/firewall_tests-e22257a8a15a7116)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 3 filtered out; finished in 0.00s

     Running tests/integration_tests.rs (target/debug/deps/integration_tests-8e7dfa9921effefa)

running 1 test
test snat_icmp4_echo_rewrite ... FAILED

failures:

---- snat_icmp4_echo_rewrite stdout ----
thread 'snat_icmp4_echo_rewrite' panicked at 'called `Result::unwrap()` on an `Err` value: StartPastEnd', /home/ry/src/opte/opte/src/engine/packet.rs:1392:51
stack backtrace:
   0: rust_begin_unwind
             at /rustc/90743e7298aca107ddaa0c202a4d3604e29bfeb6/library/std/src/panicking.rs:575:5
   1: core::panicking::panic_fmt
             at /rustc/90743e7298aca107ddaa0c202a4d3604e29bfeb6/library/core/src/panicking.rs:65:14
   2: core::result::unwrap_failed
             at /rustc/90743e7298aca107ddaa0c202a4d3604e29bfeb6/library/core/src/result.rs:1791:5
   3: core::result::Result<T,E>::unwrap
             at /rustc/90743e7298aca107ddaa0c202a4d3604e29bfeb6/library/core/src/result.rs:1113:23
   4: opte::engine::packet::Packet<opte::engine::packet::Parsed>::hdr_seg
             at /home/ry/src/opte/opte/src/engine/packet.rs:1392:17
   5: opte::engine::packet::Packet<opte::engine::packet::Parsed>::emit_new_headers
             at /home/ry/src/opte/opte/src/engine/packet.rs:1494:9
   6: opte::engine::port::Port<N>::process
             at /home/ry/src/opte/opte/src/engine/port.rs:1112:13
   7: integration_tests::snat_icmp4_echo_rewrite
             at ./tests/integration_tests.rs:990:15
   8: integration_tests::snat_icmp4_echo_rewrite::{{closure}}
             at ./tests/integration_tests.rs:956:1
   9: core::ops::function::FnOnce::call_once
             at /rustc/90743e7298aca107ddaa0c202a4d3604e29bfeb6/library/core/src/ops/function.rs:251:5
  10: core::ops::function::FnOnce::call_once
             at /rustc/90743e7298aca107ddaa0c202a4d3604e29bfeb6/library/core/src/ops/function.rs:251:5
note: Some details are omitted, run with `RUST_BACKTRACE=full` for a verbose backtrace.


failures:
    snat_icmp4_echo_rewrite

test result: FAILED. 0 passed; 1 failed; 0 ignored; 0 measured; 23 filtered out; finished in 0.05s

error: test failed, to rerun pass `--test integration_tests`
```

### With patch

```
ry@masaka:~/src/opte/oxide-vpc$ RUST_BACKTRACE=1 cargo test snat_icmp4_echo_rewrite
   Compiling opte v0.1.0 (/home/ry/src/opte/opte)
   Compiling oxide-vpc v0.1.0 (/home/ry/src/opte/oxide-vpc)
    Finished test [unoptimized + debuginfo] target(s) in 6.85s
     Running unittests src/lib.rs (target/debug/deps/oxide_vpc-09fa660a95f8ab19)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 6 filtered out; finished in 0.00s

     Running tests/firewall_tests.rs (target/debug/deps/firewall_tests-e22257a8a15a7116)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 3 filtered out; finished in 0.00s

     Running tests/integration_tests.rs (target/debug/deps/integration_tests-8e7dfa9921effefa)

running 1 test
test snat_icmp4_echo_rewrite ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 23 filtered out; finished in 0.01s
```

Fixes #349 
Fixes #351